Refactor ilm collectors
